+++ lib/install/elm.rb
@@ -12,12 +12,12 @@ copy_file "#{__dir__}/examples/elm/hello_elm.js",
           "#{Webpacker::Configuration.entry_path}/hello_elm.js"
 
 puts "Installing all elm dependencies"
-run "yarn add elm"
-run "yarn add --dev elm-hot-loader elm-webpack-loader"
+run "yarn add elm elm-webpack-loader"
+run "yarn add --dev elm-hot-loader"
 run "yarn run elm package install -- --yes"
 
 puts "Updating Webpack paths to include Elm file extension"
-insert_into_file Webpacker::Configuration.file_path, "    - .elm\n", after: /extensions:\n/
+insert_into_file Webpacker::Configuration.file_path, "      - .elm\n", after: /extensions:\n/
 
 puts "Updating elm source location"
 source_path = File.join Webpacker::Configuration.source, Webpacker::Configuration.paths.fetch(:entry, "packs")