+++ lib/install/template.rb
@@ -1,11 +1,19 @@
 # Install webpacker
+copy_file "#{__dir__}/config/webpacker.yml", "config/webpacker.yml"
+
 puts "Copying webpack core config and loaders"
 directory "#{__dir__}/config/webpack", "config/webpack"
 directory "#{__dir__}/config/loaders/core", "config/webpack/loaders"
-copy_file "#{__dir__}/config/.postcssrc.yml", ".postcssrc.yml"
 
-puts "Copying .babelrc to app root directory"
-copy_file "#{__dir__}/config/.babelrc", ".babelrc"
+unless File.exist?(".postcssrc.yml")
+  puts "Copying .postcssrc.yml to app root directory"
+  copy_file "#{__dir__}/config/.postcssrc.yml", ".postcssrc.yml"
+end
+
+unless File.exist?(".babelrc")
+  puts "Copying .babelrc to app root directory"
+  copy_file "#{__dir__}/config/.babelrc", ".babelrc"
+end
 
 puts "Creating javascript app source directory"
 directory "#{__dir__}/javascript", "#{Webpacker::Configuration.source}"